+++ src-tauri/src/lib.rs
@@ -388,10 +388,10 @@ pub fn run() {
             mpris_set_metadata,
             mpris_set_playback,
             audio::commands::audio_play,
-            audio::commands::audio_pause,
-            audio::commands::audio_resume,
-            audio::commands::audio_stop,
-            audio::commands::audio_seek,
+            audio::transport_commands::audio_pause,
+            audio::transport_commands::audio_resume,
+            audio::transport_commands::audio_stop,
+            audio::transport_commands::audio_seek,
             audio::mix_commands::audio_set_volume,
             audio::mix_commands::audio_update_replay_gain,
             audio::mix_commands::audio_set_eq,
